refs #2295: don't construct response on cancellation
On iOS, if in HTTPNSURLRequest::handleResult(), we receive an error of
type NSURLErrorCancelled, we don't construct a response object and
merely call async.send() after setting the response status to
ResponseStatus::Canceled.

Here, I've modified things to similarly do that (albeit in a brittle way
since OKHTTP handles cancelations with the onFailure() handler).

In my testing, this avoids an outright crash, but still hangs the map
rendering and the whole app on device rotate after catching an
in-progress HTTP download.